feat(projects): API mockup

+++ src/projects/projects.controller.ts
@@ -1,34 +1,47 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
-import { ProjectsService } from './projects.service';
-import { CreateProjectDto } from './dto/create-project.dto';
-import { UpdateProjectDto } from './dto/update-project.dto';
+import {
+	Controller,
+	Get,
+	Post,
+	Body,
+	Patch,
+	Param,
+	Delete,
+  Query
+} from '@nestjs/common'
+import { ProjectsService } from './projects.service'
+import { CreateProjectDto } from './dto/create-project.dto'
+import { UpdateProjectDto } from './dto/update-project.dto'
 
 @Controller('projects')
 export class ProjectsController {
-  constructor(private readonly projectsService: ProjectsService) {}
+	constructor(private readonly projectsService: ProjectsService) {}
 
-  @Post()
-  create(@Body() createProjectDto: CreateProjectDto) {
-    return this.projectsService.create(createProjectDto);
-  }
+	@Post()
+	create(@Body() createProjectDto: CreateProjectDto) {
+		return this.projectsService.create(createProjectDto)
+	}
 
-  @Get()
-  findAll() {
-    return this.projectsService.findAll();
-  }
+	@Get()
+	findAll(@Query('userId') userId?: string) {
+    if (userId) return this.projectsService.findUserProjects(userId)
+		
+    return this.projectsService.findAll()
+	}
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.projectsService.findOne(+id);
-  }
+	@Get(':id')
+	findOne(@Param('id') id: string) {
+		return this.projectsService.findOne(+id)
+	}
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProjectDto: UpdateProjectDto) {
-    return this.projectsService.update(+id, updateProjectDto);
-  }
+	@Patch(':id')
+	update(@Param('id') id: string, @Body() updateProjectDto: UpdateProjectDto) {
+		return this.projectsService.update(+id, updateProjectDto)
+	}
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.projectsService.remove(+id);
-  }
+	@Delete(':id')
+	remove(@Param('id') id: string) {
+		return this.projectsService.remove(+id)
+	}
+
+  
 }
